separate long routes from appModule
